Split up betterproto into separate modules
